GattServerV5/requirements.txt
ultralytics
paddleocr
pyspellchecker
fuzzywuzzy
python-Levenshtein
requests
python-dotenv
opencv-python-headless
dbus-python
